feat(sponsorship): add seeder

+++ database/migrations/2023_08_24_104400_create_sponsorships_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 20);
             $table->double('price', 3, 2);
-            $table->tinyInteger('hours');
+            $table->smallInteger('hours');
             $table->timestamps();
         });
     }
